DIS-37 Phase 6 + 7: extract reconciler, lifecycle-hooks, media-seed
Phase 6 — `agents/reconciler.ts`:
Lifted the periodic drift-correction loop and orphaned-session cleanup
out of AgentManager. The reconciler is a factory function taking
(pool, logger, runtime, diagnostics, sessionPrefix, getAgent,
setAgentStatus, setSystemLatestEvent) — wired up in the manager's
constructor. Single public method `reconcile()` runs both passes
(status + cleanup) and returns the records the manager needs to
re-broadcast over SSE. The two STUCK_*_TIMEOUT_S magic numbers are
now named module-level constants.

Phase 7 — small standalone helpers out of the manager:
- `agents/lifecycle-hooks.ts` — `runLifecycleHook(name, agent, logger)`.
  Repo-defined "stop" hook runner. Not session-coupled, takes an agent
  record + a logger; the manager passes `this.logger`. Adds a
  `LifecycleHookName` type so future hooks (e.g. archive, restart)
  have a discoverable enumeration point.
- `agents/media-seed.ts` — `seedInitialMedia(pool, agentId, mediaDir,
  files)` + `timestampMediaFileName`. Initial-media seeder for the
  createAgent path. Runs the writeFile + DB-INSERT loop that doesn't
  belong on the lifecycle owner.

manager.ts: 1671 → 1565 LOC. Cumulative across DIS-37: 4967 → 1565 (-68.5%).

Stale imports cleaned up: `runCommand`, `loadRepoHooks`, `writeFile`
no longer needed by manager.ts.

Phase 7 deliberately did NOT:
- Rename `AgentManager` (would churn ~18 callers for marginal naming
  honesty; the name is fine for what's left).
- Split CRUD from lifecycle (the methods are used together; splitting
  would force one consumer to import two classes).
- Retire the Phase 0 façade delegations (~270 LOC of 1-line wrappers
  for telemetry/persona-reviews/feedback/etc.). Those are cheap
  forwarders that preserve back-compat; migrating callers to the
  modules directly is a separate cleanup whenever the team wants it.

What's left in manager.ts is genuinely the agent-record lifecycle
owner: CRUD (list/get/rename), createAgent's orchestration (~270 LOC
— still the bulk, fundamentally inert/tmux split), the setup
callbacks, start/stop/archive, pin upserts, the runtime/diagnostics/
event/reconciler façades, the baseAgentSelectSql (~70 LOC), and the
small back-compat delegations.

790/798 server vitest pass (no test changes). Typecheck + e2e green.
